[REFACTOR] refactor(service): modularize error handling function
